// core/codeapps/workflow-console/src/generated/models/Faf001_bpdatavaluesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Faf001_bpdatavaluesfaf001_datatype = {
  324010000: 'Text',
  324010001: 'Number',
  324010002: 'Boolean',
  324010003: 'DateTime',
  324010004: 'JSON',
  324010005: 'Reference',
  324010006: 'Markdown'
} as const;
export type Faf001_bpdatavaluesfaf001_datatype = keyof typeof Faf001_bpdatavaluesfaf001_datatype;
export const Faf001_bpdatavaluesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Faf001_bpdatavaluesstatecode = keyof typeof Faf001_bpdatavaluesstatecode;
export const Faf001_bpdatavaluesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Faf001_bpdatavaluesstatuscode = keyof typeof Faf001_bpdatavaluesstatuscode;

export interface Faf001_bpdatavaluesBase {
  faf001_bpdatavalueid: string;
  "faf001_bpinstanceid@odata.bind": string;
  "faf001_bpstepid@odata.bind"?: string;
  "faf001_bptaskid@odata.bind"?: string;
  faf001_datakey: string;
  faf001_datatype: Faf001_bpdatavaluesfaf001_datatype;
  faf001_description?: string;
  faf001_valueboolean?: boolean;
  faf001_valuedatetime?: string;
  faf001_valuejson?: string;
  faf001_valuenumber?: number;
  faf001_valuetext?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Faf001_bpdatavaluesstatecode;
  statuscode?: Faf001_bpdatavaluesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Faf001_bpdatavalues extends Faf001_bpdatavaluesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  faf001_bpinstanceidname?: string;
  faf001_bpstepidname?: string;
  faf001_bptaskidname?: string;
  faf001_datatypename?: string;
  faf001_valuebooleanname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  faf001_bpinstanceid?: object;
  _faf001_bpinstanceid_value?: string;
  faf001_bpstepid?: object;
  _faf001_bpstepid_value?: string;
  faf001_bptaskid?: object;
  _faf001_bptaskid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}