// core/codeapps/workflow-console/src/generated/models/Faf001_bpattachmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Faf001_bpattachmentsfaf001_attachmenttype = {
  324010000: 'Document',
  324010001: 'Image',
  324010002: 'Email',
  324010003: 'Data File',
  324010004: 'Other'
} as const;
export type Faf001_bpattachmentsfaf001_attachmenttype = keyof typeof Faf001_bpattachmentsfaf001_attachmenttype;
export const Faf001_bpattachmentsfaf001_source = {
  324010000: 'Email',
  324010001: 'User Upload',
  324010002: 'SharePoint',
  324010003: 'Blob Storage',
  324010004: 'API',
  324010005: 'External System'
} as const;
export type Faf001_bpattachmentsfaf001_source = keyof typeof Faf001_bpattachmentsfaf001_source;
export const Faf001_bpattachmentsfaf001_storagetype = {
  324010000: 'OneDrive',
  324010001: 'SharePoint',
  324010002: 'Dataverse',
  324010003: 'Azure Blob Storage',
  324010004: 'Other'
} as const;
export type Faf001_bpattachmentsfaf001_storagetype = keyof typeof Faf001_bpattachmentsfaf001_storagetype;
export const Faf001_bpattachmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Faf001_bpattachmentsstatecode = keyof typeof Faf001_bpattachmentsstatecode;
export const Faf001_bpattachmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Faf001_bpattachmentsstatuscode = keyof typeof Faf001_bpattachmentsstatuscode;

export interface Faf001_bpattachmentsBase {
  faf001_attachmenttype?: Faf001_bpattachmentsfaf001_attachmenttype;
  faf001_bpattachmentid: string;
  "faf001_bpinstanceid@odata.bind"?: string;
  "faf001_bpstepid@odata.bind"?: string;
  "faf001_bptaskid@odata.bind"?: string;
  faf001_description?: string;
  faf001_externalurl?: string;
  faf001_filename: string;
  faf001_filesize?: number;
  faf001_mimetype?: string;
  faf001_receivedon?: string;
  faf001_source?: Faf001_bpattachmentsfaf001_source;
  faf001_storagetype?: Faf001_bpattachmentsfaf001_storagetype;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Faf001_bpattachmentsstatecode;
  statuscode?: Faf001_bpattachmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Faf001_bpattachments extends Faf001_bpattachmentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  faf001_attachmenttypename?: string;
  faf001_bpinstanceidname?: string;
  faf001_bpstepidname?: string;
  faf001_bptaskidname?: string;
  faf001_file?: string;
  faf001_file_name?: string;
  faf001_sourcename?: string;
  faf001_storagetypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  faf001_bpinstanceid?: object;
  _faf001_bpinstanceid_value?: string;
  faf001_bpstepid?: object;
  _faf001_bpstepid_value?: string;
  faf001_bptaskid?: object;
  _faf001_bptaskid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Faf001_bpattachmentsFileColumnName = 'faf001_file';

export type Faf001_bpattachmentsUploadColumnName = Faf001_bpattachmentsFileColumnName;